requirements.txt
asgiref==3.6.0
Django==4.1.7
django-bootstrap5==22.2
django-taggit==3.1.0
Markdown==3.4.1
psycopg2==2.9.5
psycopg2-binary==2.9.5
python-decouple==3.8
sqlparse==0.4.3
tzdata==2022.7
